streamer: replace ansi_term with enable_ansi_support
ansi_term appears to be unmaintained, and the only functionality needed
by us, enable_ansi_support(), is provided by a single-purpose crate,
enable-ansi-support. It is now defined for all platforms (if not on
windows, it is a no-op), eliminating frustrating conditional
compilation. Currently, its Result is unwrapped, which will cause a
panic on Windows versions less than 10.